requirements.txt
pydantic>=2.5.0
httpx>=0.26.0
pyyaml>=6.0.0
langgraph>=0.2.0
crewai>=0.30.0
pyautogen>=0.2.0
